// backend/src/article/enum/article-family.enum.ts
export enum ArticleFamily {
  //Textile
  Tshirt = 'Tshirt',
  Sweat = 'Sweat',
  Polo = 'Polo',
  Debardeur = 'Débardeur',
  Chemise = 'Chemise',
  Bodywarmer = 'Bodywarmer',
  Casquette = 'Casquette',
  Bonnet = 'Bonnet',
  // Veste = 'Veste',
  // Accessoire = 'Accessoire',
  //Objet
  Masque = 'Masque',
  Mug = 'Mug',
  Stylo = 'Stylo',
  Briquet = 'Briquet',
  Bracelet = 'Bracelet',
  Gobelet = 'Gobelet',
  Gourde = 'Gourde',
  Sticker = 'Sticker',
  Kakemono = 'Kakémono',
  //Pack
  Startup = 'Startup',
  Wei = 'Wei',
  Diplome = 'Diplôme',
  Campagne = 'Campagne',
  Soiree = 'Soirée',
  Sport = 'Sport',
  Festival = 'Féstival',
  Eco = 'Éco',
  Hygiene = 'Hygiène',
}